[fix] 삭제 파일 rebase - #63

+++ dateroad-domain/src/main/java/org/dateroad/place/repository/CoursePlaceRepository.java
@@ -1,6 +1,7 @@
 package org.dateroad.place.repository;
 
 import java.util.List;
+import org.dateroad.date.domain.Course;
 import org.dateroad.place.domain.CoursePlace;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
@@ -10,6 +11,5 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface CoursePlaceRepository extends JpaRepository<CoursePlace,Long> {
     @Query("SELECT SUM(p.duration) FROM CoursePlace p WHERE p.course.id = :courseId")
-    Float findTotalDurationByCourseId(@Param("courseId") Long courseId);
-
+    Integer findTotalDurationByCourseId(@Param("courseId") Long courseId);
 }
